extract repos to standalone modules
