+++ src/lib/vocab/post/post_middleware.ts
@@ -12,7 +12,7 @@ export const to_posts_middleware = (server: ApiServer): Middleware => {
 		}
 		console.log('[post_middleware] space', req.params.space_id);
 
-		const find_posts_result = await db.repos.posts.filter_by_space(req.params.space_id);
+		const find_posts_result = await db.repos.post.filter_by_space(req.params.space_id);
 		if (find_posts_result.ok) {
 			return send(res, 200, {posts: find_posts_result.value}); // TODO API types
 		} else {
@@ -35,7 +35,7 @@ export const to_create_post_middleware = (server: ApiServer): Middleware => {
 
 		// TODO take content from body & build post to pass along with it
 
-		const insert_posts_result = await db.repos.posts.insert(
+		const insert_posts_result = await db.repos.post.insert(
 			req.body.actor_id,
 			req.params.space_id,
 			req.body.content,
